Removed some pyc files. Converted the tomodachi module to use unicode so it works with Python2. TODO: Fix python2 dependency in epd.

+++ epd/epd2in7b.py
@@ -24,10 +24,10 @@
  # THE SOFTWARE.
  #
 
-import epdif
-import Image
-import ImageDraw
-import ImageFont
+from . import epdif
+from PIL import Image
+from PIL import ImageDraw
+from PIL import ImageFont
 import RPi.GPIO as GPIO
 
 # Display resolution
@@ -259,7 +259,7 @@ class EPD:
             self.send_data(self.lut_wb[count])
 
     def get_frame_buffer(self, image):
-        buf = [0xFF] * (self.width * self.height / 8)
+        buf = [0xFF] * int(self.width * self.height / 8)
         # Set buffer to value of Python Imaging Library image.
         # Image must be in mode 1.
         image_monocolor = image.convert('1')
